Add weather bonus multiplier

+++ src/main/java/decorator/monster/MonsterDecorator.java
@@ -72,7 +72,7 @@ public abstract class MonsterDecorator implements CodeAMon {
     }
 
     @Override
-    public void setWeatherBonus(int weatherBonus) {
+    public void setWeatherBonus(double weatherBonus) {
         this.weatherBonus = weatherBonus;
     }
